Move from registration execution to flow execution
