+++ components/org.wso2.carbon.identity.api.server.flow.execution/org.wso2.carbon.identity.api.server.flow.execution.v1/src/main/java/org/wso2/carbon/identity/api/server/flow/execution/v1/impl/RegistrationApiServiceImpl.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.api.server.flow.execution.v1.impl;
+
+import org.wso2.carbon.identity.api.server.flow.execution.v1.FlowExecutionRequest;
+import org.wso2.carbon.identity.api.server.flow.execution.v1.RegistrationApiService;
+import org.wso2.carbon.identity.api.server.flow.execution.v1.core.FlowExecutionServiceCore;
+import org.wso2.carbon.identity.api.server.flow.execution.v1.factories.FlowExecutionServiceFactory;
+
+import javax.ws.rs.core.Response;
+
+/**
+ * Implementation of the Registration API.
+ */
+public class RegistrationApiServiceImpl implements RegistrationApiService {
+
+    private final FlowExecutionServiceCore flowExecutionServiceCore;
+
+    public RegistrationApiServiceImpl() {
+
+        this.flowExecutionServiceCore = FlowExecutionServiceFactory.getFlowExecutionService();
+    }
+
+    @Override
+    public Response registrationExecutePost(FlowExecutionRequest flowExecutionRequest) {
+
+        return Response.ok()
+                .entity(flowExecutionServiceCore.processFlowExecution((flowExecutionRequest)))
+                .build();
+    }
+}
